mod: Add socket buffer reservation
Inspired by the iSCSI soft initiator in icl_soft.c, do some additional
setup to reserve buffer space for the socket once handed off to the
kernel.  The max payload size sent by nbd-server is too large to be
reserved, so create a tunable for the client to configure an upper
limit on this size.