mod: Reserve more sockbuf space if needed
In some cases the kernel may reduce our socket buffer space and disable
autosizing.  This can lead to the sb_hiwat mark being lower than
sb_lowat, which blocks us in sbwait() forever.  Check for this
pathological behavior and work around it.

There should be a better way to handle this, but this fixes a real issue
for now.