mod: Let send buffer drain more
nbd_conn_send spends() a good amount of time in sbwait() when the send
buffer is full.  Rather than waiting for enough space to send once, take
a hint from icl_soft.c and wait for a fraction of the high water mark if
it is greater.  This means we may sleep for longer each time but will
have to sleep less frequently, reducing overall time in sbwait().