Large write support
Lift the limit requiring maxpayload to fit in the default socket buffer
size or in a single TLS record in the case of TLS connections.  Now we
reserve socket buffers as large as the system allows, up to the protocol
recommended 2^25 bytes or advertised maxpayload, and write requests are
divided into multiple mbuf chains as required to fit the constraints of
the socket buffers and TLS records.  Each chain is sent before building
the next chain, minimizing memory overhead.