mod: Rework waiting for soft disconnect completion
We avoid doing a wakeup every time the inflight queue is emptied during
normal operation by observing the state transition to SOFT_DISCONNECTING
and checking for inflight bios in the receiver thread.

The receiver is allowed to continue during SOFT_DISCONNECTING until all
inflight bios are completed or an error degrades the state to
HARD_DISCONNECTING.  Then the receiver can perform the state transition
to HARD_DISCONNECTING before exiting.  The sender need only wait for the
receiver to be done.