ci: Raise lo0 mtu for profiles
GCE images set ifconfig_DEFAULT="SYNCDHCP mtu 1460" which unfortunately
applies to the loopback interface lo0 as well.  Raise this back to the
default 16384 for profiles so we can have it low for the coverage to
catch regressions during the functional tests but have it high for
more reasonable behavior during profiling.