Persist devices with failed connections
Allow devices to persist when all connections have failed.  The device
can be recovered by the `scale` command to resume normal operation.  The
`disconnect` command will now wait for the device to be destroyed.
Likewise, module unloading now waits for all devices to be destroyed.

In-flight requests for failed connections are not requeued.  This is
left for future work.