lib: Fix width vs precision confusion in format strings
"%*s" specifies width from an argument, while "%.*s" specifies precision.
Only precision truncates the string to the given length.

Fixes devices having garbage at the end of their name and/or description
when the server provides those fields, which in turn causes geom to
reject the dumpconf string when the garbage is not valid ASCII.