mod: Do not conflate stripesize and block size
GEOM does not have a way to report preferred block size, and NBD does
not have a way to convey stripe size.  Stop conflating the two.

A user can create a gnop with the desired stripe size and sector size
for now. Future work could add optional parameters to gnbd connect to
force a particular sector size and stripe size.