mod: Hide the irrelevant port for Unix-domain sockets
We can't easily remove the default (or otherwise) port from the gctl_req
when we find we are connecting to a Unix-domain socket.  The port is
irrelevant, so replace it with an empty string in the kernel softc.
That way it will not clutter up the GEOM::ident string for Unix-domain
socket connections.  Leave the ':' that usually delimits the port, so it
doesn't end up looking like one big path.