mod: Avoid socantsendmore
Use soshutdown(SHUT_WR) instead of socantsendmore(). The former informs
the protocol that we are done sending on the socket, while the latter is
for protocols to implement it.  The new unix/stream code does not work
with socantsendmore().  Instead, it implements SHUT_WR using an internal
uipc_cantrcvmore() on the receiving end.