mod: Fix error on reconnect retry
A command like `gnbd reconnect -r 5 nbd0` (e.g. executed by devd on
disconnect) would always fail if the first attempt to connect failed.
But, the failure would only appear to happen after connection to the
server should be possible.

In fact, the reconnect does eventually succeed.  However, the prior
failed connection attempts would have already set an error on the gctl
request, ultimately resulting in failure.

Add a flag to nbd_client_connect to prevent setting an error result for
the gctl request when we are in a retry loop.  The only way we exit the
retry loop is on success, we don't want to keep the previous errors.