mod: Fix refcounting after disconnect
Move initialization of the bio refcount to its first insertion to the
main queue.  Do not release a reference when moving inflight bios back
to the main queue.  This ensures that the refcount stays accurate when a
connection fails and an inflight write with references to the bio gets
requeued.

We can't expect the network stack to have finished with a bio before we
requeue it, and when the stack does finally free its mbufs we need the
bio to still have a reference so it is not delivered prematurely.

For the same reason, we must not deliver bios in g_nbd_drain_queue if
the network stack is still holding references.  Use the existing
inflight delivery mechanism to handle this.

Reserve a cookie for the main queue, for debugging purposes.