ci: Pipe distsets from fetch to tar
GCE images are supposed to come with the sources.  However, many of the
published images are broken and missing sources.  And, the GCE image may
have installed updates to the kernel and sources at boot and not
rebooted.

We need the installed kernel and source to match what is currently
running.  We cannot reboot.  So, we must continue to unconditionally
fetch and extract the kernel, kernel-dbg, and src distsets.

One slight optimization we can make is to pipe directly from fetch to
tar instead of hitting the filesystem.