ci: Fix kmod debug symbols install on 15.0
Installing the kmod debug symbols started failing on 15.0 GCE images
because of a missing directory.  This could be fixed in the Makefile,
but it seems like more of a FreeBSD build system problem.  Just rely on
the user to make the directory if it is missing, and do exactly this in
the CI.